Refactor previous types file into new dir
